5730 Move FFIFileResult to foreign-bridge module, update Javadoc
