+++ java/foreign-bridge/src/main/java/sleeper/foreign/FFIFileResult.java
@@ -13,21 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.datafusion;
+package sleeper.foreign;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import jnr.ffi.Struct;
 
 /**
- * The compaction output data that the native code will populate.
+ * The file output data that the native code will populate.
+ *
+ * <strong>THIS IS A C COMPATIBLE FFI STRUCT!</strong> If you updated this struct (field ordering, types, etc.),
+ * you MUST update the corresponding Rust definition in rust/sleeper_df/src/objects.rs. The order and types of
+ * the fields must match exactly.
  */
 @SuppressWarnings(value = {"checkstyle:membername", "checkstyle:parametername"})
 @SuppressFBWarnings(value = {"URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD"})
-public class DataFusionCompactionResult extends Struct {
+public class FFIFileResult extends Struct {
     public final Struct.size_t rows_read = new Struct.size_t();
     public final Struct.size_t rows_written = new Struct.size_t();
 
-    public DataFusionCompactionResult(jnr.ffi.Runtime runtime) {
+    public FFIFileResult(jnr.ffi.Runtime runtime) {
         super(runtime);
     }
 }
